Python Bindings: Restructuring, Documentation
Restructure the package to make it cleaner, export the KNN graph function as an
Extra so installing faiss is not required, export command line capabilities of
the graph creation function. Add README to long_description of setup. Reformat
bindings. Add additional options to graph maker that allow controlling the hnsw
index. Skip the test when data is not present.